chore: use verdaccio memory plugin to speed up e2e tests
This also removes writing of `.npmrc` files when possible. It is only
needed for the auth token on publish.
The registry for the installs is now set via an env variable, as this
works most reliable in different npm versions.
Npm 6/Node 12 for example does not pickup project `.npmrc` files.